requirements.txt
requests==2.3.0
websocket-client==0.14.1
